Added new 1D-CNN examples for feature extraction, separation of AE networks for Synthetic and RGC Datasets, Bugfixing stuff in prepare_training files

+++ 3_Python/src_data/merge_rgc_onoff_fzj.py
@@ -29,7 +29,7 @@ if __name__ == "__main__":
         cluster_dict.append(folder.split('_')[-1][:-4])
 
     frames_in = np.array(frames_in, dtype=np.float16)
-    frames_cl = np.array(frames_cl, dtype=np.uint8)
+    frames_cl = np.array(frames_cl, dtype=np.int16)
 
     # --- Saving the data
     savemat(f'../data/{datetime.now().strftime("%Y-%m-%d")}_rgc_onoff_fzj_Merged.mat',
